sea-tool-backend/requirements.txt
-i https://pypi.org/simple
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==3.2.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-registration==0.5.6
django==2.2.7
djangorestframework-jwt==1.11.0
djangorestframework==3.10.3
drf-nested-routers==0.91
drf-writable-nested==0.5.4
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itypes==1.1.0
jinja2==2.10.3
jsonschema==3.2.0
markdown==3.1.1
markupsafe==1.1.1
more-itertools==7.2.0
packaging==20.4
pillow==7.1.2
pyjwt==1.7.1
pyparsing==2.4.5
pyrsistent==0.15.5
python-dotenv==0.10.3
pytz==2019.3
pyyaml==5.1.2
requests==2.22.0
ruamel.yaml.clib==0.2.0
ruamel.yaml==0.16.5
six==1.13.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.7
zipp==0.6.0
